Added disk caching of asts to reduce RAM usage and speedup reruns.
New function loadOrParseAST() which checks if the ast on disk is up-to-date. If so it reloads the serialized AST from disk, otherwise it parses it again from the source file and afterwards serialize it to disk.

So far, serialized ASTs are saved with extension .ast at the places where the respective source file (.cpp) is stored.